Revise content node's brief to display a substring of the text
`Content` node's (textual) brief field was initially designed
to allow developers display a minute of their content on the grid,
for easier review and accessibility.
It used to accept a separate text and could be really cumbersome;
Specially when devs intended to display a large part
or all of the main text.

New approach replaces that with a brief-length which defines
length of a substring from beginning of the main content text,
to be shown as brief on the grid node.

`-1` can be set as the value to display full text on the grid.

This commit also allows content nodes to be resized on the grid.

CAUTION!
Note also that, although deprecated textual briefs are still supported as legacy
and displayed properly by both the console and official HTML-JS runtime,
they will be merged into the beginning of the main text of the content node
upon inspection, and updated in the project file on the first save.
It does not change the content nodes that are not opened in the inspector
or not (auto-) updated on inspection.