Remove `package.json` file from HTML-JS runtime
This manifest file is of no use at the time; so is discarded.

We do not use `npm` to build our HTML-JS runtime anymore.
And because any change in Arrow regarding running nodes
shall immediately reflect in the runtime code as well;
the package now has the same versioning as the Arrow itself.