Revise and simplify console handling of macros and node-variable state tracking
This commit simplifies the console code,
mainly by moving macro special treatments to the node's console module,
and revising the way we track and scope played nodes and variable states.

It also fixes some inconsistencies and is part of an effort
to make console and runtime behaviors as similar as possible.
In other words standardization of what we expect from a runtime in general.